# artifacts/api-server/src/routes/tests.ts
import { Router } from "express";
import { db, questionsTable, testSessionsTable, usersTable } from "@workspace/db";
import { eq, desc, and, inArray } from "drizzle-orm";
import { v4 as uuidv4 } from "uuid";
import { requireAuth } from "../middlewares/auth";
import { logger } from "../lib/logger";

const router = Router();
const TEST_QUESTIONS = 25;
const TEST_DURATION_SECONDS = 8 * 60; // 8 minutes
const PASS_THRESHOLD = 0.6;

router.post("/start", requireAuth, async (req, res) => {
  try {
    const { userId } = (req as typeof req & { user: { userId: number } }).user;
    const { mode = "timed" } = req.body;

    const allQuestions = await db.select().from(questionsTable).where(eq(questionsTable.status, "published"));
    if (allQuestions.length < TEST_QUESTIONS) {
      res.status(400).json({ error: "Not enough questions in the bank" });
      return;
    }

    // Distribute: ~40% easy, ~40% medium, ~20% hard
    const easy = shuffle(allQuestions.filter(q => q.difficulty === "easy"));
    const medium = shuffle(allQuestions.filter(q => q.difficulty === "medium"));
    const hard = shuffle(allQuestions.filter(q => q.difficulty === "hard"));

    const selectedEasy = easy.slice(0, Math.min(10, easy.length));
    const selectedMedium = medium.slice(0, Math.min(10, medium.length));
    const selectedHard = hard.slice(0, Math.min(5, hard.length));

    let selected = shuffle([...selectedEasy, ...selectedMedium, ...selectedHard]);
    if (selected.length < TEST_QUESTIONS) {
      const remaining = allQuestions.filter(q => !selected.includes(q));
      selected = [...selected, ...shuffle(remaining)].slice(0, TEST_QUESTIONS);
    } else {
      selected = selected.slice(0, TEST_QUESTIONS);
    }

    const now = new Date();
    const expiresAt = new Date(now.getTime() + TEST_DURATION_SECONDS * 1000);
    const sessionId = uuidv4();

    await db.insert(testSessionsTable).values({
      sessionId,
      userId,
      mode,
      questionIds: selected.map(q => q.id),
      status: "in_progress",
      startedAt: now,
      expiresAt,
    });

    res.status(201).json({
      sessionId,
      questions: selected.map(q => ({
        ...q,
        options: q.options as string[],
        createdAt: q.createdAt?.toISOString() ?? now.toISOString(),
      })),
      startedAt: now.toISOString(),
      expiresAt: expiresAt.toISOString(),
      durationSeconds: TEST_DURATION_SECONDS,
      mode,
    });
  } catch (err) {
    logger.error({ err }, "Start test error");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/:sessionId/submit", requireAuth, async (req, res) => {
  try {
    const { userId } = (req as typeof req & { user: { userId: number } }).user;
    const { sessionId } = req.params;
    const { answers } = req.body;

    const [session] = await db.select().from(testSessionsTable)
      .where(and(eq(testSessionsTable.sessionId, sessionId), eq(testSessionsTable.userId, userId)))
      .limit(1);

    if (!session) {
      res.status(404).json({ error: "Session not found" });
      return;
    }
    if (session.status !== "in_progress") {
      res.status(400).json({ error: "Session already completed" });
      return;
    }

    const now = new Date();
    const expired = now > session.expiresAt;
    const timeTaken = Math.min(
      Math.floor((now.getTime() - session.startedAt.getTime()) / 1000),
      TEST_DURATION_SECONDS
    );

    const questionIds = session.questionIds as number[];
    const questions = await db.select().from(questionsTable).where(inArray(questionsTable.id, questionIds));
    const questionMap = new Map(questions.map(q => [q.id, q]));

    const answersMap = new Map<number, number>(
      (answers ?? []).map((a: { questionId: number; selectedAnswer: number }) => [a.questionId, a.selectedAnswer])
    );

    let score = 0;
    const answerDetails = questionIds.map(qId => {
      const q = questionMap.get(qId);
      if (!q) return null;
      const selected = answersMap.get(qId) ?? -1;
      const isCorrect = selected === q.correctAnswer;
      if (isCorrect) score++;
      return {
        questionId: qId,
        selectedAnswer: selected,
        correctAnswer: q.correctAnswer,
        isCorrect,
        explanation: q.explanation,
      };
    }).filter(Boolean);

    const total = questionIds.length;
    const percentage = (score / total) * 100;
    const passed = percentage >= PASS_THRESHOLD * 100;

    await db.update(testSessionsTable).set({
      answers: answerDetails as typeof testSessionsTable.$inferSelect['answers'],
      score,
      total,
      percentage,
      passed,
      timeTaken,
      status: expired ? "expired" : "completed",
      completedAt: now,
    }).where(eq(testSessionsTable.sessionId, sessionId));

    // Update user XP and stats
    const xpGain = passed ? 100 : 30;
    const [user] = await db.select().from(usersTable).where(eq(usersTable.id, userId)).limit(1);
    if (user) {
      const newXp = (user.xp ?? 0) + xpGain;
      const newTotalTests = (user.totalTests ?? 0) + 1;
      const currentPassRate = user.passRate ?? 0;
      const newPassRate = ((currentPassRate * (newTotalTests - 1)) + (passed ? 100 : 0)) / newTotalTests;
      const newLevel = Math.floor(newXp / 500) + 1;
      await db.update(usersTable).set({
        xp: newXp,
        totalTests: newTotalTests,
        passRate: newPassRate,
        level: newLevel,
        lastActiveAt: now,
      }).where(eq(usersTable.id, userId));
    }

    res.json({
      sessionId,
      score,
      total,
      percentage,
      passed,
      timeTaken,
      completedAt: now.toISOString(),
      mode: session.mode,
      answers: answerDetails,
    });
  } catch (err) {
    logger.error({ err }, "Submit test error");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/history", requireAuth, async (req, res) => {
  try {
    const { userId } = (req as typeof req & { user: { userId: number } }).user;
    const { limit = "20", offset = "0" } = req.query as Record<string, string>;
    const sessions = await db.select().from(testSessionsTable)
      .where(and(eq(testSessionsTable.userId, userId), eq(testSessionsTable.status, "completed")))
      .orderBy(desc(testSessionsTable.completedAt))
      .limit(parseInt(limit))
      .offset(parseInt(offset));

    res.json(sessions.map(s => ({
      sessionId: s.sessionId,
      score: s.score ?? 0,
      total: s.total ?? 25,
      percentage: s.percentage ?? 0,
      passed: s.passed ?? false,
      timeTaken: s.timeTaken ?? 0,
      completedAt: s.completedAt?.toISOString() ?? s.startedAt.toISOString(),
      mode: s.mode,
    })));
  } catch (err) {
    logger.error({ err }, "Get test history error");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/:sessionId", requireAuth, async (req, res) => {
  try {
    const { userId } = (req as typeof req & { user: { userId: number } }).user;
    const { sessionId } = req.params;
    const [session] = await db.select().from(testSessionsTable)
      .where(and(eq(testSessionsTable.sessionId, sessionId), eq(testSessionsTable.userId, userId)))
      .limit(1);

    if (!session) {
      res.status(404).json({ error: "Test result not found" });
      return;
    }

    res.json({
      sessionId: session.sessionId,
      score: session.score ?? 0,
      total: session.total ?? 25,
      percentage: session.percentage ?? 0,
      passed: session.passed ?? false,
      timeTaken: session.timeTaken ?? 0,
      completedAt: session.completedAt?.toISOString() ?? session.startedAt.toISOString(),
      mode: session.mode,
      answers: session.answers ?? [],
    });
  } catch (err) {
    logger.error({ err }, "Get test result error");
    res.status(500).json({ error: "Internal server error" });
  }
});

function shuffle<T>(arr: T[]): T[] {
  const a = [...arr];
  for (let i = a.length - 1; i > 0; i--) {
    const j = Math.floor(Math.random() * (i + 1));
    [a[i], a[j]] = [a[j], a[i]];
  }
  return a;
}

export default router;