# lib/api-zod/src/generated/types/answerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VID Master Zimbabwe API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerInput {
  questionId: number;
  selectedAnswer: number;
}