# lib/api-zod/src/generated/types/testSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VID Master Zimbabwe API
 * OpenAPI spec version: 0.1.0
 */
import type { AnswerInput } from './answerInput';

export interface TestSubmission {
  answers: AnswerInput[];
}